fixed klippekot

+++ src/Application/Controller/Controller.java
@@ -137,7 +137,7 @@ public class Controller {
         return rl;
     }
 
-    public int brugKlippekort(Klippekort klippekort, int klipPris){
+    public double brugKlippekort(Klippekort klippekort, double klipPris){
         klippekort.brugKlip(klipPris);
         return klippekort.getAntalKlip();
     }
